[fixing feedback] put all filtering into a different util file
